Rename prometheus_utils to metrics_utils
