+++ cas/scheduler/BUILD
@@ -27,7 +27,7 @@ rust_library(
         ":platform_property_manager",
         ":worker",
         "//util:error",
-        "//util:prometheus_utils",
+        "//util:metrics_utils",
         "@crate_index//:tokio",
     ],
 )
@@ -82,7 +82,7 @@ rust_library(
         "//config",
         "//util:common",
         "//util:error",
-        "//util:prometheus_utils",
+        "//util:metrics_utils",
         "@crate_index//:parking_lot",
         "@crate_index//:futures",
         "@crate_index//:hashbrown",
@@ -216,7 +216,7 @@ rust_library(
         "//cas/store",
         "//config",
         "//util:error",
-        "//util:prometheus_utils",
+        "//util:metrics_utils",
         "@crate_index//:futures",
         "@crate_index//:tokio",
     ],
@@ -234,7 +234,7 @@ rust_library(
         ":platform_property_manager",
         "//proto",
         "//util:error",
-        "//util:prometheus_utils",
+        "//util:metrics_utils",
         "@crate_index//:tokio",
         "@crate_index//:uuid",
     ],
@@ -252,7 +252,7 @@ rust_library(
         "//proto",
         "//util:common",
         "//util:error",
-        "//util:prometheus_utils",
+        "//util:metrics_utils",
         "@crate_index//:prost",
         "@crate_index//:prost-types",
         "@crate_index//:sha2",
