+++ cas/scheduler/worker.rs
@@ -21,8 +21,8 @@ use uuid::Uuid;
 
 use action_messages::ActionInfo;
 use error::{make_err, make_input_err, Code, Error, ResultExt};
+use metrics_utils::{CollectorState, CounterWithTime, FuncCounterWrapper, MetricsComponent};
 use platform_property_manager::{PlatformProperties, PlatformPropertyValue};
-use prometheus_utils::{CollectorState, CounterWithTime, FuncCounterWrapper, MetricsComponent};
 use proto::com::github::allada::turbo_cache::remote_execution::{
     update_for_worker, ConnectionResult, StartExecute, UpdateForWorker,
 };
