+++ cas/store/BUILD
@@ -36,7 +36,7 @@ rust_library(
         ":verify_store",
         "//config",
         "//util:error",
-        "//util:prometheus_utils",
+        "//util:metrics_utils",
         "@crate_index//:futures",
     ],
 )
@@ -50,7 +50,7 @@ rust_library(
         "//util:buf_channel",
         "//util:common",
         "//util:error",
-        "//util:prometheus_utils",
+        "//util:metrics_utils",
         "@crate_index//:bytes",
         "@crate_index//:futures",
         "@crate_index//:serde",
@@ -119,7 +119,7 @@ rust_library(
         "//util:common",
         "//util:error",
         "//util:evicting_map",
-        "//util:prometheus_utils",
+        "//util:metrics_utils",
         "@crate_index//:bytes",
     ],
 )
@@ -135,7 +135,7 @@ rust_library(
         "//util:buf_channel",
         "//util:common",
         "//util:error",
-        "//util:prometheus_utils",
+        "//util:metrics_utils",
         "@crate_index//:hex",
         "@crate_index//:sha2",
         "@crate_index//:tokio",
@@ -242,7 +242,7 @@ rust_library(
         "//util:common",
         "//util:error",
         "//util:evicting_map",
-        "//util:prometheus_utils",
+        "//util:metrics_utils",
         "@crate_index//:async-lock",
         "@crate_index//:bytes",
         "@crate_index//:filetime",
