+++ cas/worker/BUILD
@@ -15,7 +15,7 @@ rust_library(
         "//proto",
         "//util:common",
         "//util:error",
-        "//util:prometheus_utils",
+        "//util:metrics_utils",
         "@crate_index//:async-lock",
         "@crate_index//:futures",
         "@crate_index//:tokio",
@@ -39,7 +39,7 @@ rust_library(
         "//proto",
         "//util:common",
         "//util:error",
-        "//util:prometheus_utils",
+        "//util:metrics_utils",
         "@crate_index//:bytes",
         "@crate_index//:filetime",
         "@crate_index//:futures",
