+++ cas/worker/running_actions_manager.rs
@@ -28,8 +28,8 @@ use bytes::{Bytes, BytesMut};
 use filetime::{set_file_mtime, FileTime};
 use futures::future::{try_join, try_join3, try_join_all, BoxFuture, Future, FutureExt, TryFutureExt};
 use futures::stream::{FuturesUnordered, StreamExt, TryStreamExt};
+use metrics_utils::{AsyncCounterWrapper, CollectorState, CounterWithTime, MetricsComponent};
 use parking_lot::Mutex;
-use prometheus_utils::{AsyncCounterWrapper, CollectorState, CounterWithTime, MetricsComponent};
 use prost::Message;
 use relative_path::RelativePath;
 use tokio::io::{AsyncReadExt, AsyncSeekExt};
