+++ gencargo/action_messages/Cargo.toml
@@ -29,4 +29,4 @@ platform_property_manager = { workspace = true }
 proto = { workspace = true }
 common = { workspace = true }
 error = { workspace = true }
-prometheus_utils = { workspace = true }
+metrics_utils = { workspace = true }
