+++ gencargo/cas/Cargo.toml
@@ -46,4 +46,4 @@ local_worker = { workspace = true }
 config = { workspace = true }
 common = { workspace = true }
 error = { workspace = true }
-prometheus_utils = { workspace = true }
+metrics_utils = { workspace = true }
