+++ gencargo/filesystem_store/Cargo.toml
@@ -37,4 +37,4 @@ buf_channel = { workspace = true }
 common = { workspace = true }
 error = { workspace = true }
 evicting_map = { workspace = true }
-prometheus_utils = { workspace = true }
+metrics_utils = { workspace = true }
