+++ gencargo/local_worker/Cargo.toml
@@ -36,4 +36,4 @@ config = { workspace = true }
 proto = { workspace = true }
 common = { workspace = true }
 error = { workspace = true }
-prometheus_utils = { workspace = true }
+metrics_utils = { workspace = true }
