+++ gencargo/metrics_utils/Cargo.toml
@@ -4,7 +4,7 @@
 # Do not edit this file directly.
 
 [package]
-name = "scheduler"
+name = "metrics_utils"
 version = "0.0.0"
 edition = "2021"
 autobins = false
@@ -13,18 +13,14 @@ autotests = false
 autobenches = false
 
 [lib]
-name = "scheduler"
-path = "../../cas/scheduler/scheduler.rs"
+name = "metrics_utils"
+path = "../../util/metrics_utils.rs"
 # TODO(allada) We should support doctests.
 doctest = false
 
 [dependencies]
-async-trait = { workspace = true }
-tokio = { workspace = true }
+futures = { workspace = true }
+prometheus-client = { workspace = true }
 
 # Local libraries.
-action_messages = { workspace = true }
-platform_property_manager = { workspace = true }
-worker = { workspace = true }
-error = { workspace = true }
-metrics_utils = { workspace = true }
+
