+++ gencargo/running_actions_manager/Cargo.toml
@@ -42,4 +42,4 @@ config = { workspace = true }
 proto = { workspace = true }
 common = { workspace = true }
 error = { workspace = true }
-prometheus_utils = { workspace = true }
+metrics_utils = { workspace = true }
