+++ gencargo/traits/Cargo.toml
@@ -28,4 +28,4 @@ serde = { workspace = true }
 buf_channel = { workspace = true }
 common = { workspace = true }
 error = { workspace = true }
-prometheus_utils = { workspace = true }
+metrics_utils = { workspace = true }
