+++ gencargo/verify_store/Cargo.toml
@@ -30,4 +30,4 @@ config = { workspace = true }
 buf_channel = { workspace = true }
 common = { workspace = true }
 error = { workspace = true }
-prometheus_utils = { workspace = true }
+metrics_utils = { workspace = true }
