+++ gencargo/worker/Cargo.toml
@@ -27,4 +27,4 @@ action_messages = { workspace = true }
 platform_property_manager = { workspace = true }
 proto = { workspace = true }
 error = { workspace = true }
-prometheus_utils = { workspace = true }
+metrics_utils = { workspace = true }
